pyproject.toml
[project]
name = "aichhoernchen"
description = "A Münsterhack 2025 Project"
version = "0.0.1"
license = "MIT"
authors = [
    { name = "Paul-B98", email = "115164840+Paul-B98@users.noreply.github.com" }  # TODO
]
maintainers = [
    { name = "Paul-B98", email = "115164840+Paul-B98@users.noreply.github.com" }  # TODO
]
readme = "README.md"
keywords = ["Python Lib"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "openai-agents>=0.3.2",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "smolagents[openai]>=1.22.0",
    "strawberry-graphql-django>=0.65.1",
]

[project.urls]
homepage = "https://github.com/Just-another-Muensterhack/aichhoernchen-backend"
repository = "https://github.com/Just-another-Muensterhack/aichhoernchen-backend"
changelog = "https://github.com/Just-another-Muensterhack/aichhoernchen-backend/releases"
issues = "https://github.com/Just-another-Muensterhack/aichhoernchen-backend/issues"
documentation = "https://github.com/Just-another-Muensterhack/aichhoernchen-backend/wiki"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aichhoernchen"]

[dependency-groups]
dev = [
    "jupyterlab>=4.4.3",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.13",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-material>=9.6.14",
    "mkdocs-section-index>=0.3.10",
    "mkdocstrings-python>=1.16.12",
    "pymdown-extensions>=10.15",
]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
filterwarnings = ["ignore::DeprecationWarning"]
pythonpath = ["src"]
cache_dir = "/workspaces/.cache/pytest"
addopts = [
    "--cov",
    "--cov-branch",
    "--cov=src",
    "--cov-report=xml",
    "--import-mode=importlib",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict_optional = true
exclude = [
    "docs/*",
    "tests/*",
]
